Simplifies client URL handling and validation
Simplifies the client initialization process by enforcing a trailing slash in the base URL. This change removes redundant port handling logic.

Also, it throws an exception when calling a function on an invalid simulation, improving error handling and preventing unexpected behavior.

+++ src/nominalpy/simulation/simulation.py
@@ -140,6 +140,12 @@ class Simulation(Context):
         simulation can no longer be used after this has been called.
         """
 
+        # Throw exception if the simulation is not valid
+        if not self.is_valid():
+            raise NominalException(
+                "Failed to call function on an invalid or deleted simulation."
+            )
+
         # Dispose of the simulation and reset the state
         await self.__client.delete(f"{self.__id}")
         self.__reset()
